+++ ChangeLog.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 - Added :openai module support for assistant run api (...)
 - Added :openai module support for assistant run steps api (...)
 
+- Added function `ipc/clone` to clone clients
+
 
 
 ## [1.12.61] - 2025-11-16