+++ ChangeLog.md
@@ -14,7 +14,8 @@ All notable changes to this project will be documented in this file.
 - Added :openai module support for assistant run api (...)
 - Added :openai module support for assistant run steps api (...)
 
-- Added function `ipc/clone` to clone clients
+- Added function `ipc/clone` to clone IPC clients
+- Added IPC message expiration